requirements.txt
Flask==1.1.2
flask-paginate
Flask-Mail
gunicorn
google-cloud-datastore
google-cloud-storage
